requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

boto3==1.34.125
numpy==1.26.4
pandas==2.2.2
plotly==5.22.0
PyYAML==6.0.1
requests==2.31.0
streamlit==1.34.0
streamlit-authenticator==0.3.2
streamlit-extras==0.4.2
openpyxl==3.1.2